requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    gup requirements.txt
#
--find-links file:///home/oliver/.cache/pip-wheelhouse

assetbuilder==1.0.0       # via -r requirements.in
astunparse==1.6.3         # via piglet-templates
attrs==19.3.0             # via piglet-templates
babel==2.8.0              # via fresco-i18n
diskcache==4.1.0          # via -r requirements.in
embrace-sql==1.0          # via -r requirements.in
embrace==2.2.0            # via embrace-sql
https://hg.sr.ht/~olly/fresco-i18n/archive/a3672a8dd6dc.tar.gz#egg=fresco-i18n  # via -r requirements.in
fresco-static==0.2        # via -r requirements.in
fresco-template==0.3.1    # via -r requirements.in
fresco==1.0.1             # via -r requirements.in, fresco-i18n, fresco-static, fresco-template
gevent==20.6.2            # via -r requirements.in
greenlet==0.4.16          # via gevent
gunicorn==20.0.4          # via -r requirements.in
itsdangerous==1.1.0       # via assetbuilder
markupsafe==1.1.1         # via piglet-templates
parsley==1.3              # via piglet-templates
pathlib==1.0.1            # via assetbuilder
piglet-templates==1.0.0   # via -r requirements.in
portalocker==1.7.0        # via assetbuilder
psycopg2==2.8.5           # via -r requirements.in
pytz==2020.1              # via -r requirements.in, babel
six==1.15.0               # via astunparse
sqlparse==0.3.1           # via embrace, yoyo-migrations
tabulate==0.8.7           # via yoyo-migrations
toml==0.10.1              # via -r requirements.in
wheel==0.34.2             # via astunparse
wrapt==1.12.1             # via embrace
yoyo-migrations==7.1.2    # via -r requirements.in
zope.event==4.4           # via gevent
zope.interface==5.1.0     # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
